Move Backoff to libkb
